move test only update functions to _test.go files
